+++ R/data.R
@@ -1,0 +1,27 @@
+#' Toy example
+#'
+#' A data set containing patient IDs, event types, event times, and
+#' gender of 100 patients.
+#'
+#' @format {A data frame with 104 rows and 4 columns
+#' \describe{
+#'    \item{PTID}{ID number of patients}
+#'    \item{EvTp}{Event Types: SHK as Shock, CHF as Congestive Heart
+#'          Failure, REMI as Recurrent Myocardial Infarction, DTH as
+#'          Death; and N as No event}
+#'    \item{EvTm}{Event Time (day)}
+#'    \item{sex}{Gender of patients, M as Male, F as Female}
+#'    }
+#'}
+#'
+#'@source {It is a generated example based on ASSENT-3:
+#' \url{https://pubmed.ncbi.nlm.nih.gov/21570513}}
+#'
+#'@section References:
+#'  Armstrong P. W., Westerhout C. M., Van de Werf F., Califf R. M., Welsh R. C.,
+#'  Wilcox R. G., Bakal J. A. (2011) Refining clinical trial composite outcomes:
+#'  an application to the Assessment of the Safety and Efficacy of a New
+#'  Thrombolytic-3 (ASSENT-3) trial. \var{American Heart Journal}. \bold{161}(5)
+#'  848-854.
+#'
+"toyexample"