+++ NEWS.md
@@ -1,3 +1,6 @@
 # wcep 1.0.0
 
-* Initial CRAN submission.
+* Added a `NEWS.md` file to track changes to the package.
+
+## Comments
+This is the first version of this package. There is no news.